backend/services/service-engagement-tracker/requirements.txt
# Core dependencies
fastapi==0.121.0
uvicorn[standard]==0.38.0
pydantic==2.12.3
pydantic-settings==2.7.0
python-dotenv==1.2.1

# Database
sqlalchemy>=2.0.36  # Python 3.13 compatible
psycopg[binary]>=3.1.0  # Python 3.13 compatible
alembic>=1.13.1

# Data processing & ML
pandas>=2.2.0  # Compatible with Python 3.13
numpy>=1.26.0
scikit-learn>=1.4.0
joblib>=1.3.2

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Utilities
faker>=22.0.0  # For synthetic data generation

# Security
python-multipart==0.0.20
email-validator==2.3.0

# HTTP client
httpx==0.28.1

# Optional: Remove if not needed
# sentry-sdk==2.43.0  # Error tracking
